change the file stucture and fix some bugs
