+++ Frontend/dairy-farm/src/App.jsx
@@ -1,30 +1,29 @@
 import "./App.css";
-import NavBar from "./components/NavBar";
+import NavBar from "./components/Sales/NavBar";
 import "@fontsource/roboto/400.css";
 import CssBaseline from "@mui/material/CssBaseline";
-import ProductGrid from "./components/ProductGrid";
-import CustomerSignUp from "./pages/CustomerSignUp";
-import CustomerSignIn from "./pages/CustomerSignIn";
+import ProductGrid from "./components/Sales/ProductGrid";
+import CustomerSignUp from "./pages/Sales/CustomerSignUp";
+import CustomerSignIn from "./pages/Sales/CustomerSignIn";
 import { RouterProvider } from "react-router-dom";
 import router from "./routes";
 import { AuthProvider } from "./contexts/AuthContext";
-import ProductDetails from "./components/ProductDetails";
-import CheckOut from "./components/CheckOut";
-import SalesDashboard from "./pages/SalesDashboard";
+import ProductDetails from "./components/Sales/ProductDetails";
+import CheckOut from "./components/Sales/CheckOut";
+import SalesDashboard from "./pages/Sales/SalesDashboard";
 import MilkingDashboard from "./pages/Milking/MilkingDashboard";
-import ProductList from "./components/ProductList";
+import ProductList from "./components/Sales/ProductList";
 
 function App() {
   return (
     <>
-      {/* <AuthProvider>
+      <AuthProvider>
         <CssBaseline />
         <RouterProvider router={router} />
       </AuthProvider>
-      <CssBaseline /> */}
+      <CssBaseline />
 
-      <MilkingDashboard />
-
+      {/* <MilkingDashboard /> */}
     </>
   );
 }
