+++ Frontend/dairy-farm/src/components/Sales/CustomerList.jsx
@@ -5,16 +5,14 @@ import TableCell from "@mui/material/TableCell";
 import TableHead from "@mui/material/TableHead";
 import TableRow from "@mui/material/TableRow";
 import SalesTitle from "./SalesTitle";
-import usePurcahse from "../hooks/usePurcahse";
 import Button from "@mui/material/Button";
-import useProducts from "../hooks/useProducts";
-import useGameQueryStore from "../store";
+import useGameQueryStore from "../../store";
 import { Link } from "react-router-dom";
-import publishService from "../services/publishService";
+import publishService from "../../services/Sales/publishService";
+import useCustomers from "../../hooks/useCustomers";
+import approveService from "../../services/Sales/approveService";
 import Box from "@mui/material/Box";
 import Skeleton from "@mui/material/Skeleton";
-import productService from "../services/productService";
-import DeleteIcon from "@mui/icons-material/Delete";
 
 // Generate Order Data
 function createData(id, date, name, shipTo, paymentMethod, amount) {
@@ -68,8 +66,8 @@ function preventDefault(event) {
   event.preventDefault();
 }
 
-const ProductList = () => {
-  const { data, error, isLoading, refetch } = useProducts();
+const CustomerList = () => {
+  const { data, error, isLoading, refetch } = useCustomers();
 
   const SetSelectedProductUpdate = useGameQueryStore(
     (s) => s.SetSelectedProductUpdate
@@ -85,11 +83,11 @@ const ProductList = () => {
     SetSelectedProductUpdate(id);
   };
 
-  const handlePublish = (id, publish) => {
-    const data = { publish: publish };
+  const handleApprove = (id, approvel) => {
+    const Approvedata = { approvel: approvel };
 
-    publishService
-      .Publish(id, data)
+    approveService
+      .Approve(id, Approvedata)
       .then((res) => {
         console.log(res.data);
         console.log("success");
@@ -100,81 +98,55 @@ const ProductList = () => {
       });
   };
 
-  const handleDelete = (id) => {
-    productService
-      .Delete(id)
-      .then((res) => {
-        console.log(res.data);
-        refetch();
-      })
-      .catch((err) => {
-        console.log(err.message);
-      });
-  };
-
   return (
-    <React.Fragment>
-      <SalesTitle>Recent Products</SalesTitle>
+    <>
+      <SalesTitle>Customers</SalesTitle>
       <Table size="small">
         <TableHead>
           <TableRow>
             <TableCell>Name</TableCell>
-            <TableCell>Quantity</TableCell>
-            <TableCell>Category</TableCell>
-            <TableCell align="right">Unit Price</TableCell>
-            <TableCell>Update</TableCell>
-            <TableCell>Publish</TableCell>
-            <TableCell>Delete</TableCell>
+            <TableCell>address</TableCell>
+            <TableCell>Phone</TableCell>
+            <TableCell>license No </TableCell>
+            <TableCell align="right">Rep</TableCell>
+            <TableCell>email</TableCell>
+            <TableCell>Approve</TableCell>
+            <TableCell>Cancel</TableCell>
           </TableRow>
         </TableHead>
         <TableBody>
-          {data?.map((product) => (
-            <TableRow key={product._id}>
-              <TableCell>{product.name}</TableCell>
-              <TableCell>{product.quantity}</TableCell>
-              <TableCell>{product.category}</TableCell>
-              <TableCell align="right">{`$${product.price}`}</TableCell>
+          {data?.map((customer) => (
+            <TableRow key={customer._id}>
+              <TableCell>{customer.name}</TableCell>
+              <TableCell>{customer.address}</TableCell>
+              <TableCell>{customer.phone}</TableCell>
+              <TableCell>{customer.licenseNo}</TableCell>
+              <TableCell align="right">{customer.rep}</TableCell>
+              <TableCell>{customer.email}</TableCell>
+
               <TableCell>
-                <Link to="/productupdate">
+                {customer.approvel == false ? (
                   <Button
-                    onClick={() => handleUpdate(product._id)}
-                    variant="contained"
-                    size="medium"
-                    color="success"
-                  >
-                    Update
-                  </Button>
-                </Link>
-              </TableCell>
-              <TableCell>
-                {product.publish == false ? (
-                  <Button
-                    onClick={() => handlePublish(product._id, true)}
+                    onClick={() => handleApprove(customer._id, true)}
                     variant="outlined"
                     size="medium"
                     color="error"
                   >
-                    Unpublish
+                    Approve
                   </Button>
                 ) : (
                   <Button
-                    onClick={() => handlePublish(product._id, false)}
+                    onClick={() => handleApprove(customer._id, false)}
                     variant="outlined"
                     size="medium"
                     color="error"
                   >
-                    Publish
+                    Cancel
                   </Button>
                 )}
               </TableCell>
               <TableCell>
-                <Button
-                  onClick={() => handleDelete(product._id)}
-                  variant="outlined"
-                  size="medium"
-                  color="error"
-                  startIcon={<DeleteIcon />}
-                >
+                <Button variant="outlined" size="medium" color="error">
                   Delete
                 </Button>
               </TableCell>
@@ -182,6 +154,7 @@ const ProductList = () => {
           ))}
         </TableBody>
       </Table>
+
       {isLoading ? (
         <Box sx={{ width: 1100 }}>
           <Skeleton sx={{ height: 80, marginTop: -1 }} />
@@ -194,8 +167,8 @@ const ProductList = () => {
       <Link color="primary" href="#" onClick={preventDefault} sx={{ mt: 3 }}>
         See more orders
       </Link>
-    </React.Fragment>
+    </>
   );
 };
 
-export default ProductList;
+export default CustomerList;
