+++ Frontend/dairy-farm/src/components/Sales/InputField.jsx
