+++ Frontend/dairy-farm/src/components/Sales/NavBar.jsx
