+++ Frontend/dairy-farm/src/components/Sales/SalesNavBar.jsx
