--- Frontend/dairy-farm/src/pages/Layout.jsx
@@ -1,14 +1,0 @@
-import React from 'react'
-import NavBar from '../components/NavBar';
-import { Outlet } from 'react-router-dom';
-
-const Layout = () => {
-  return (
-    <>
-        <NavBar/>
-        <Outlet/>
-    </>
-  )
-}
-
-export default Layout;
