+++ Frontend/dairy-farm/src/pages/Sales/Layout.jsx
@@ -1,0 +1,14 @@
+import React from "react";
+import NavBar from "../../components/Sales/NavBar";
+import { Outlet } from "react-router-dom";
+
+const Layout = () => {
+  return (
+    <>
+      <NavBar />
+      <Outlet />
+    </>
+  );
+};
+
+export default Layout;
