+++ Frontend/dairy-farm/src/routes.jsx
@@ -1,13 +1,13 @@
 import { createBrowserRouter } from "react-router-dom";
-import Layout from "./pages/Layout";
-import HomePage from "./pages/HomePage";
-import CustomerSignUp from "./pages/CustomerSignUp";
-import CustomerSignIn from "./pages/CustomerSignIn";
-import ProtectedRoute from "./components/ProtectedRoute";
-import ProductDetails from "./components/ProductDetails";
-import CheckOut from "./components/CheckOut";
-import ProductUpdate from "./pages/ProductUpdate";
-import SalesDashboard from "./pages/SalesDashboard";
+import Layout from "./pages/Sales/Layout";
+import HomePage from "./pages/Sales/HomePage";
+import CustomerSignUp from "./pages/Sales/CustomerSignUp";
+import CustomerSignIn from "./pages/Sales/CustomerSignIn";
+import ProtectedRoute from "./components/Sales/ProtectedRoute";
+import ProductDetails from "./components/Sales/ProductDetails";
+import CheckOut from "./components/Sales/CheckOut";
+import ProductUpdate from "./pages/Sales/ProductUpdate";
+import SalesDashboard from "./pages/Sales/SalesDashboard";
 
 const router = createBrowserRouter([
   {
@@ -26,7 +26,7 @@ const router = createBrowserRouter([
         path: "/productdetails",
         element: (
           <ProtectedRoute>
-            <ProductDetails/>
+            <ProductDetails />
           </ProtectedRoute>
         ),
       },
@@ -34,10 +34,10 @@ const router = createBrowserRouter([
         path: "/checkout",
         element: (
           <ProtectedRoute>
-            <CheckOut/>
+            <CheckOut />
           </ProtectedRoute>
-        )
-      }
+        ),
+      },
     ],
   },
   {
@@ -50,12 +50,12 @@ const router = createBrowserRouter([
   },
   {
     path: "/productupdate",
-    element: <ProductUpdate/>
+    element: <ProductUpdate />,
   },
   {
     path: "/salesdashboard",
-    element: <SalesDashboard/>
-  }
+    element: <SalesDashboard />,
+  },
 ]);
 
 export default router;
